+++ lib/active_type/change_association.rb
@@ -1,0 +1,26 @@
+module ActiveType
+
+  module ChangeAssociation
+
+    extend ActiveSupport::Concern
+
+    module ClassMethods
+
+      def change_association(association_name, new_scope, new_options = nil)
+        if (existing_association = self.reflect_on_association(association_name))
+          if new_scope.is_a?(Hash)
+            new_options = new_scope
+            new_scope = existing_association.scope
+          end
+          original_options = existing_association.options
+          public_send(existing_association.macro, association_name, new_scope, original_options.merge(new_options))
+        else
+          raise ArgumentError, "unrecognized association `#{association_name}`"
+        end
+      end
+
+    end
+
+  end
+
+end