+++ lib/active_type/record.rb
@@ -1,6 +1,7 @@
 require 'active_type/virtual_attributes'
 require 'active_type/record_extension'
 require 'active_type/nested_attributes'
+require 'active_type/change_association'
 
 module ActiveType
 
@@ -11,6 +12,7 @@ module ActiveType
     include VirtualAttributes
     include NestedAttributes
     include RecordExtension
+    include ChangeAssociation
 
   end